azure-certifications-study-guide/AI-projects/project-01-document-processing/src/requirements.txt
azure-ai-documentintelligence==1.0.0b1
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
